+++ CHANGELOG.md
@@ -28,6 +28,9 @@
 *   New method Spree.formatMoney(amount, currency) is available to javascript in
     the admin.
 
+*   Fix an issue where updating a user in the admin without specifying roles in
+    the params would clear the user's roles.
+
 *   Deprecations
 
     * `cache_key_for_taxons` helper has been deprecated in favour of `cache [I18n.locale, @taxons]`